+++ packages/playground/src/index.tsx
@@ -1,5 +1,5 @@
 import * as ReactDOM from 'react-dom';
-import App from './page/App';
+import App from './App';
 import './style/fonts.css';
 import './style/tailwind.css';
 import './style/global.css';
